// apps/backend/src/modules/atividades/exercise-image-type.enum.ts
export enum ExerciseImageType {
  MOBILIDADE_GERAL = 'MOBILIDADE_GERAL',
  MOBILIDADE_LOMBAR = 'MOBILIDADE_LOMBAR',
  CONTROLE_CERVICAL = 'CONTROLE_CERVICAL',
  OMBRO_MANGUITO = 'OMBRO_MANGUITO',
  JOELHO_AGACHAMENTO = 'JOELHO_AGACHAMENTO',
  QUADRIL_GLUTEOS = 'QUADRIL_GLUTEOS',
  TORNOZELO_EQUILIBRIO = 'TORNOZELO_EQUILIBRIO',
  PUNHO_PREENSAO = 'PUNHO_PREENSAO',
  MOBILIDADE_LOMBAR_GATO_CAMELO = 'MOBILIDADE_LOMBAR_GATO_CAMELO',
  PONTE_CURTA = 'PONTE_CURTA',
  CONTROLE_CERVICAL_PROFUNDO = 'CONTROLE_CERVICAL_PROFUNDO',
  ELEVACAO_ASSISTIDA_OMBRO = 'ELEVACAO_ASSISTIDA_OMBRO',
  AGACHAMENTO_PARCIAL_ASSISTIDO = 'AGACHAMENTO_PARCIAL_ASSISTIDO',
  ABDUCAO_QUADRIL_DECUBITO_LATERAL = 'ABDUCAO_QUADRIL_DECUBITO_LATERAL',
  EQUILIBRIO_BIPODAL_TRANSFERENCIA_PESO = 'EQUILIBRIO_BIPODAL_TRANSFERENCIA_PESO',
  PREENSAO_MANUAL_BOLA_MACIA = 'PREENSAO_MANUAL_BOLA_MACIA',
  MOBILIDADE_TORACICA_ROTACAO_SENTADA = 'MOBILIDADE_TORACICA_ROTACAO_SENTADA',
  RETRACAO_ESCAPULAR_SENTADA = 'RETRACAO_ESCAPULAR_SENTADA',
  ISOMETRIA_ROTACAO_EXTERNA_OMBRO = 'ISOMETRIA_ROTACAO_EXTERNA_OMBRO',
  EXTENSAO_JOELHO_SENTADO = 'EXTENSAO_JOELHO_SENTADO',
  SENTAR_LEVANTAR_CONTROLADO = 'SENTAR_LEVANTAR_CONTROLADO',
  ALONGAMENTO_FLEXORES_QUADRIL_MEIO_AJOELHADO = 'ALONGAMENTO_FLEXORES_QUADRIL_MEIO_AJOELHADO',
  MARCHA_ESTACIONARIA_APOIO = 'MARCHA_ESTACIONARIA_APOIO',
  MOBILIDADE_PUNHO_FLEXAO_EXTENSAO = 'MOBILIDADE_PUNHO_FLEXAO_EXTENSAO',
  ALONGAMENTO_CERVICAL_LATERAL_ASSISTIDO = 'ALONGAMENTO_CERVICAL_LATERAL_ASSISTIDO',
  DESLIZAMENTO_NEURAL_MEDIANO = 'DESLIZAMENTO_NEURAL_MEDIANO',
}

export const EXERCISE_IMAGE_TYPES = Object.values(ExerciseImageType);